Refactor cycle autonomy plumbing
