+++ brain/systems/cycles/memory.py
@@ -1,0 +1,388 @@
+"""Durable Cycle memory, revisions, snapshots, and evaluations."""
+from __future__ import annotations
+
+from sqlalchemy import func, select
+
+from brain.platform.db.models.cycle import (
+    Cycle,
+    CycleGuidance,
+    CycleOutputTarget,
+    CycleRevision,
+    CycleRun,
+    CycleRunEvaluation,
+)
+from brain.systems.cycles.common import (
+    CYCLE_LEDGER_OUTPUT_TARGET_TYPE,
+    THREAD_OUTPUT_TARGET_TYPE,
+    actor_id,
+    actor_type,
+    creator_payload,
+    json_dict,
+    string_or_none,
+    validate_nonempty_trimmed,
+)
+from brain.systems.cycles.serializers import (
+    serialize_cycle_guidance,
+    serialize_cycle_output_target,
+    serialize_cycle_revision,
+)
+
+
+async def async_record_cycle_revision(
+    session,
+    cycle: Cycle,
+    *,
+    source_type: str = "system",
+    source_id: str | None = None,
+    rationale: str | None = None,
+) -> CycleRevision:
+    await session.execute(select(Cycle.id).where(Cycle.id == cycle.id).with_for_update())
+    result = await session.execute(
+        select(func.coalesce(func.max(CycleRevision.revision_number), 0)).where(
+            CycleRevision.cycle_id == cycle.id
+        )
+    )
+    current_revision = int(result.scalar_one() or 0)
+    revision = CycleRevision(
+        cycle_id=cycle.id,
+        revision_number=current_revision + 1,
+        source_type=actor_type(source_type),
+        source_id=actor_id(source_id),
+        rationale=(rationale or "").strip() or None,
+        name=cycle.name,
+        prompt=cycle.prompt,
+        schedule_expr=cycle.schedule_expr,
+        timezone=cycle.timezone,
+        enabled=cycle.enabled,
+        model_override=cycle.model_override,
+        thinking_override=cycle.thinking_override,
+        target_idea_id=cycle.target_idea_id,
+        context_policy={
+            "workspace_id": string_or_none(cycle.org_id),
+            "owner_user_id": string_or_none(cycle.user_id),
+            **creator_payload(cycle),
+        },
+    )
+    session.add(revision)
+    await session.flush()
+    return revision
+
+
+async def async_add_cycle_guidance(
+    session,
+    cycle: Cycle,
+    *,
+    guidance: str,
+    source_type: str = "user",
+    source_id: str | None = None,
+    rationale: str | None = None,
+    revision_id: int | None = None,
+) -> CycleGuidance:
+    row = CycleGuidance(
+        cycle_id=cycle.id,
+        revision_id=revision_id,
+        source_type=actor_type(source_type),
+        source_id=actor_id(source_id),
+        guidance=validate_nonempty_trimmed(guidance, "guidance"),
+        rationale=(rationale or "").strip() or None,
+        is_active=True,
+    )
+    session.add(row)
+    await session.flush()
+    return row
+
+
+async def async_add_cycle_output_target(
+    session,
+    cycle: Cycle,
+    *,
+    target_type: str,
+    target_id: str | None = None,
+    label: str | None = None,
+    config: dict | None = None,
+    source_type: str = "user",
+    source_id: str | None = None,
+    rationale: str | None = None,
+    revision_id: int | None = None,
+) -> CycleOutputTarget:
+    row = CycleOutputTarget(
+        cycle_id=cycle.id,
+        revision_id=revision_id,
+        target_type=validate_nonempty_trimmed(target_type, "target_type"),
+        target_id=(str(target_id).strip() if target_id is not None else None) or None,
+        label=(label or "").strip() or None,
+        config=json_dict(config),
+        source_type=actor_type(source_type),
+        source_id=actor_id(source_id),
+        rationale=(rationale or "").strip() or None,
+        is_active=True,
+    )
+    session.add(row)
+    await session.flush()
+    return row
+
+
+async def async_remove_cycle_output_target(
+    session,
+    cycle: Cycle,
+    *,
+    target_id: int,
+    source_type: str = "user",
+    source_id: str | None = None,
+    rationale: str | None = None,
+    revision_id: int | None = None,
+) -> CycleOutputTarget | None:
+    row = await session.get(CycleOutputTarget, target_id)
+    if row is None or row.cycle_id != cycle.id:
+        return None
+    row.is_active = False
+    row.revision_id = revision_id
+    row.source_type = actor_type(source_type)
+    row.source_id = actor_id(source_id)
+    row.rationale = (rationale or "").strip() or row.rationale
+    await session.flush()
+    return row
+
+
+async def async_prepare_cycle_run_memory_snapshot(session, cycle: Cycle, run: CycleRun) -> None:
+    revision = await _async_latest_cycle_revision(session, cycle.id)
+    guidance_rows = await _async_active_cycle_guidance(session, cycle.id)
+    target_rows = await _async_active_cycle_output_targets(session, cycle.id)
+
+    if revision is not None:
+        run.revision_id = revision.id
+
+    output_targets = [serialize_cycle_output_target(target) for target in target_rows]
+    _ensure_default_output_targets(cycle, output_targets)
+
+    run.guidance_snapshot = [serialize_cycle_guidance(row) for row in guidance_rows]
+    run.output_targets_snapshot = output_targets
+    run.context_snapshot = {
+        "revision": serialize_cycle_revision(revision),
+        "workspace_id": string_or_none(cycle.org_id),
+        "owner_user_id": string_or_none(cycle.user_id),
+        **creator_payload(cycle),
+    }
+
+
+def append_cycle_run_output_target_snapshot(
+    run: CycleRun,
+    *,
+    target_type: str,
+    target_id: str,
+    label: str,
+    config: dict | None = None,
+    source_type: str = "system",
+    rationale: str | None = None,
+) -> None:
+    output_targets = list(run.output_targets_snapshot or [])
+    if any(
+        isinstance(target, dict)
+        and target.get("target_type") == target_type
+        and target.get("target_id") == target_id
+        for target in output_targets
+    ):
+        return
+    output_targets.append(
+        {
+            "target_type": target_type,
+            "target_id": target_id,
+            "label": label,
+            "config": json_dict(config),
+            "source_type": source_type,
+            "rationale": rationale,
+            "is_active": True,
+        }
+    )
+    run.output_targets_snapshot = output_targets
+
+
+def finalize_cycle_run(
+    run: CycleRun,
+    cycle: Cycle,
+    *,
+    status: str,
+    error: str | None = None,
+    skip_reason: str | None = None,
+    session=None,
+) -> None:
+    from datetime import datetime, timezone
+
+    now = datetime.now(timezone.utc)
+    run.status = status
+    run.completed_at = now
+    run.error = error
+    run.skip_reason = skip_reason
+    cycle.last_run_at = now
+    cycle.last_status = status
+    cycle.last_error = error
+    record_cycle_run_evaluation(
+        session,
+        run,
+        cycle,
+        status=status,
+        error=error,
+        skip_reason=skip_reason,
+    )
+
+
+def finalize_stale_cycle_run(
+    run: CycleRun,
+    cycle: Cycle | None,
+    *,
+    status: str,
+    error: str | None = None,
+    skip_reason: str | None = None,
+    session=None,
+) -> None:
+    from datetime import datetime, timezone
+
+    now = datetime.now(timezone.utc)
+    run.status = status
+    run.completed_at = now
+    run.error = error
+    run.skip_reason = skip_reason
+    if cycle is None:
+        return
+    record_cycle_run_evaluation(
+        session,
+        run,
+        cycle,
+        status=status,
+        error=error,
+        skip_reason=skip_reason,
+        evaluator_type="recovery",
+    )
+    cycle_last_run_at = _aware_utc(cycle.last_run_at)
+    scheduled_for = _aware_utc(run.scheduled_for)
+    if cycle_last_run_at is None or (
+        scheduled_for is not None and scheduled_for >= cycle_last_run_at
+    ):
+        cycle.last_run_at = now
+        cycle.last_status = status
+        cycle.last_error = error
+
+
+def record_cycle_run_evaluation(
+    session,
+    run: CycleRun,
+    cycle: Cycle,
+    *,
+    status: str,
+    error: str | None = None,
+    skip_reason: str | None = None,
+    evaluator_type: str = "system",
+    evaluator_id: str | None = None,
+) -> None:
+    summary = cycle_run_evaluation_summary(
+        status=status,
+        error=error,
+        skip_reason=skip_reason,
+    )
+    run.self_review_summary = summary
+    if session is None or not hasattr(session, "add") or run.id is None:
+        return
+    session.add(
+        CycleRunEvaluation(
+            cycle_id=cycle.id,
+            cycle_run_id=run.id,
+            evaluator_type=evaluator_type,
+            evaluator_id=actor_id(evaluator_id),
+            summary=summary,
+            score=1 if status == "completed" else 0 if status == "failed" else None,
+            details={
+                "status": status,
+                "error": error,
+                "skip_reason": skip_reason,
+                "agent_run_id": run.run_id,
+                "idea_id": string_or_none(run.idea_id),
+            },
+        )
+    )
+
+
+async def _async_latest_cycle_revision(session, cycle_id: int) -> CycleRevision | None:
+    result = await session.scalars(
+        select(CycleRevision)
+        .where(CycleRevision.cycle_id == cycle_id)
+        .order_by(CycleRevision.revision_number.desc(), CycleRevision.id.desc())
+        .limit(1)
+    )
+    return result.first()
+
+
+async def _async_active_cycle_guidance(session, cycle_id: int) -> list[CycleGuidance]:
+    result = await session.scalars(
+        select(CycleGuidance)
+        .where(CycleGuidance.cycle_id == cycle_id, CycleGuidance.is_active.is_(True))
+        .order_by(CycleGuidance.created_at.asc(), CycleGuidance.id.asc())
+        .limit(25)
+    )
+    return list(result.all())
+
+
+async def _async_active_cycle_output_targets(session, cycle_id: int) -> list[CycleOutputTarget]:
+    result = await session.scalars(
+        select(CycleOutputTarget)
+        .where(CycleOutputTarget.cycle_id == cycle_id, CycleOutputTarget.is_active.is_(True))
+        .order_by(CycleOutputTarget.created_at.asc(), CycleOutputTarget.id.asc())
+        .limit(25)
+    )
+    return list(result.all())
+
+
+def _ensure_default_output_targets(cycle: Cycle, output_targets: list[dict]) -> None:
+    if not any(target.get("target_type") == CYCLE_LEDGER_OUTPUT_TARGET_TYPE for target in output_targets):
+        output_targets.insert(
+            0,
+            {
+                "target_type": CYCLE_LEDGER_OUTPUT_TARGET_TYPE,
+                "target_id": str(cycle.id),
+                "label": "Cycle ledger",
+                "config": {},
+                "source_type": "system",
+                "rationale": "Implicit durable Cycle memory target.",
+                "is_active": True,
+            },
+        )
+    if cycle.target_idea_id and not any(
+        target.get("target_type") == THREAD_OUTPUT_TARGET_TYPE
+        and target.get("target_id") == str(cycle.target_idea_id)
+        for target in output_targets
+    ):
+        output_targets.append(
+            {
+                "target_type": THREAD_OUTPUT_TARGET_TYPE,
+                "target_id": str(cycle.target_idea_id),
+                "label": "Cycle thread",
+                "config": {},
+                "source_type": "system",
+                "rationale": "Implicit display thread target.",
+                "is_active": True,
+            }
+        )
+
+
+def cycle_run_evaluation_summary(
+    *,
+    status: str,
+    error: str | None = None,
+    skip_reason: str | None = None,
+) -> str:
+    if status == "completed":
+        return "Cycle run completed and was recorded in the Cycle ledger."
+    if status == "failed":
+        detail = error or "unknown failure"
+        return f"Cycle run failed and was recorded in the Cycle ledger: {detail}"
+    if status == "skipped":
+        detail = skip_reason or "unknown skip reason"
+        return f"Cycle run was skipped and recorded in the Cycle ledger: {detail}"
+    return f"Cycle run reached status {status} and was recorded in the Cycle ledger."
+
+
+def _aware_utc(value):
+    from datetime import timezone
+
+    if value is not None and value.tzinfo is None:
+        return value.replace(tzinfo=timezone.utc)
+    return value
