+++ brain/systems/runs/tool_catalog/handlers/cycles.py
@@ -2,8 +2,63 @@
 
 from __future__ import annotations
 
-from brain.systems.runs.tool_catalog.handlers.common import *
+from dataclasses import dataclass
+from typing import Any, Awaitable, Callable
+
+from sqlalchemy import and_, or_, select
+
+from brain.platform.db.models.cycle import Cycle
+from brain.platform.db.models.idea import Idea
+from brain.platform.db.models.org import User
 from brain.platform.db.repositories.unit_of_work import UnitOfWork
+from brain.systems.cycles.commands import (
+    CycleMutationActor,
+    async_add_guidance_to_cycle,
+    async_add_output_target_to_cycle,
+    async_create_cycle,
+    async_delete_cycle,
+    async_remove_output_target_from_cycle,
+    async_update_cycle,
+    cycle_change_event,
+)
+from brain.systems.cycles.events import publish_cycle_change
+from brain.systems.cycles.serializers import (
+    serialize_cycle,
+    serialize_cycle_guidance,
+    serialize_cycle_output_target,
+)
+from brain.systems.cycles.service import async_run_cycle_now
+from brain.systems.runs.tool_catalog.handlers.common import *
+
+
+@dataclass(frozen=True)
+class ManageCycleArgs:
+    action: str
+    id: int | None = None
+    name: str | None = None
+    prompt: str | None = None
+    schedule_expr: str | None = None
+    run_at: str | None = None
+    timezone: str | None = None
+    enabled: bool | None = None
+    model_override: str | None = None
+    thinking_override: str | None = None
+    target_idea_id: str | None = None
+    guidance: str | None = None
+    rationale: str | None = None
+    output_target_type: str | None = None
+    output_target_id: str | None = None
+    output_target_label: str | None = None
+    output_target_config: dict | None = None
+
+
+@dataclass(frozen=True)
+class ManageCycleContext:
+    args: ManageCycleArgs
+    actor: CycleMutationActor
+
+
+CycleAction = Callable[[ManageCycleContext], Awaitable[dict[str, Any]]]
 
 
 async def _handle_manage_cycle(
@@ -18,9 +73,7 @@ async def _handle_manage_cycle(
     enabled: bool | None = None,
     model_override: str | None = None,
     thinking_override: str | None = None,
-    execution_mode: str | None = None,
     target_idea_id: str | None = None,
-    reopen_archived: bool | None = None,
     guidance: str | None = None,
     rationale: str | None = None,
     output_target_type: str | None = None,
@@ -40,9 +93,7 @@ async def _handle_manage_cycle(
         enabled=enabled,
         model_override=model_override,
         thinking_override=thinking_override,
-        execution_mode=execution_mode,
         target_idea_id=target_idea_id,
-        reopen_archived=reopen_archived,
         guidance=guidance,
         rationale=rationale,
         output_target_type=output_target_type,
@@ -64,9 +115,7 @@ async def _handle_manage_cycle_async(
     enabled: bool | None = None,
     model_override: str | None = None,
     thinking_override: str | None = None,
-    execution_mode: str | None = None,
     target_idea_id: str | None = None,
-    reopen_archived: bool | None = None,
     guidance: str | None = None,
     rationale: str | None = None,
     output_target_type: str | None = None,
@@ -74,442 +123,277 @@ async def _handle_manage_cycle_async(
     output_target_label: str | None = None,
     output_target_config: dict | None = None,
 ) -> str:
-    action = str(action or "").strip().lower()
-    if action in {"help", "schema"}:
+    normalized_action = str(action or "").strip().lower()
+    if normalized_action in {"help", "schema"}:
         return _manage_tool_guide("manage_cycle", operation)
 
-    from sqlalchemy import and_, or_, select
-
-    from brain.systems.cycles.service import (
-        build_one_time_schedule_expr,
-        compute_next_run_at,
-        cycle_defaults,
-        REUSABLE_THREAD_EXECUTION_MODE,
-        async_add_cycle_guidance,
-        async_add_cycle_output_target,
-        async_record_cycle_revision,
-        async_remove_cycle_output_target,
-        async_run_cycle_now,
-        serialize_cycle,
-        serialize_cycle_guidance,
-        serialize_cycle_output_target,
-        validate_nonempty_trimmed,
-        validate_execution_mode,
-        validate_schedule_expr,
-        validate_thinking_override,
-        validate_timezone_name,
-    )
-    from brain.systems.cycles.events import publish_cycle_change
-    from brain.platform.db.models.cycle import Cycle
-    from brain.platform.db.models.idea import Idea
-    from brain.platform.db.models.org import User
-    from datetime import timezone as dt_timezone
+    handler = CYCLE_ACTIONS.get(normalized_action)
+    if handler is None:
+        return json.dumps({"error": f"Unknown action: {normalized_action}"})
 
     user_id = getattr(_agent_context, "user_id", None)
-    org_id = getattr(_agent_context, "org_id", None)
     if not user_id:
         return json.dumps({"error": "manage_cycle requires a user-scoped cortex run"})
 
-    def _optional_text(value: str | None) -> str | None:
-        if value is None:
-            return None
-        text = str(value).strip()
-        return text or None
-
-    def _optional_update_text(value: str | None) -> str | None:
-        # Tool-call optional string fields often arrive as empty strings. Treat blanks
-        # as omitted for update-only fields so a schedule update is not derailed by
-        # an unrelated default like run_at="" or name="".
-        return _optional_text(value)
-
-    def _cycle_scope():
-        scope = [Cycle.deleted_at.is_(None)]
-        if org_id:
-            org_user_ids = select(User.id).where(User.org_id == org_id)
-            scope.append(
-                or_(
-                    Cycle.org_id == org_id,
-                    and_(Cycle.org_id.is_(None), Cycle.user_id.in_(org_user_ids)),
-                )
-            )
-        else:
-            scope.append(Cycle.user_id == user_id)
-        return scope
-
-    def _idea_scope(idea_id: str):
-        scope = [Idea.id == idea_id]
-        if org_id:
-            org_user_ids = select(User.id).where(User.org_id == org_id)
-            scope.append(
-                or_(
-                    Idea.org_id == org_id,
-                    and_(Idea.org_id.is_(None), Idea.user_id.in_(org_user_ids)),
-                )
-            )
-        else:
-            scope.append(Idea.user_id == user_id)
-        return scope
+    args = ManageCycleArgs(
+        action=normalized_action,
+        id=id,
+        name=name,
+        prompt=prompt,
+        schedule_expr=schedule_expr,
+        run_at=run_at,
+        timezone=timezone,
+        enabled=enabled,
+        model_override=model_override,
+        thinking_override=thinking_override,
+        target_idea_id=target_idea_id,
+        guidance=guidance,
+        rationale=rationale,
+        output_target_type=output_target_type,
+        output_target_id=output_target_id,
+        output_target_label=output_target_label,
+        output_target_config=output_target_config,
+    )
+    context = ManageCycleContext(args=args, actor=_tool_actor(str(user_id)))
 
     try:
-        if action == "list":
-            async with UnitOfWork() as uow:
-                stmt = (
-                    select(Cycle)
-                    .where(*_cycle_scope())
-                    .order_by(Cycle.created_at.desc())
-                )
-                result = await uow.session.scalars(stmt)
-                cycles = [serialize_cycle(cycle) for cycle in result.all()]
-            return json.dumps({"cycles": cycles}, default=str)
-        elif action == "create":
-            create_name = _optional_text(name)
-            create_prompt = _optional_text(prompt)
-            create_timezone = _optional_text(timezone)
-            create_run_at = _optional_text(run_at)
-            create_schedule_expr = _optional_text(schedule_expr)
-            if not create_name or not create_prompt or not create_timezone or (not create_schedule_expr and not create_run_at):
-                return json.dumps({"error": "create requires: name, prompt, timezone, and schedule_expr or run_at"})
-            tz_name = validate_timezone_name(create_timezone)
-            expr = (
-                build_one_time_schedule_expr(create_run_at, tz_name)
-                if create_run_at
-                else validate_schedule_expr(create_schedule_expr or "", tz_name)
+        payload = await handler(context)
+        return json.dumps(payload, default=str)
+    except ValueError as exc:
+        return json.dumps({"error": str(exc)})
+    except Exception as exc:
+        logger.exception("manage_cycle failed: %s", exc)
+        if _is_missing_cycle_schema_error(exc):
+            return json.dumps(_cycle_schema_missing_payload(exc))
+        return json.dumps({"error": str(exc)})
+
+
+async def _action_list(ctx: ManageCycleContext) -> dict[str, Any]:
+    async with UnitOfWork() as uow:
+        result = await uow.session.scalars(
+            select(Cycle)
+            .where(*_cycle_scope(ctx.actor))
+            .order_by(Cycle.created_at.desc())
+        )
+        return {"cycles": [serialize_cycle(cycle) for cycle in result.all()]}
+
+
+async def _action_create(ctx: ManageCycleContext) -> dict[str, Any]:
+    args = ctx.args
+    async with UnitOfWork() as uow:
+        await _validate_target_idea(uow.session, args.target_idea_id, ctx.actor)
+        cycle = await async_create_cycle(
+            uow.session,
+            actor=ctx.actor,
+            name=_optional_text(args.name),
+            prompt=_optional_text(args.prompt),
+            timezone_name=_optional_text(args.timezone),
+            schedule_expr=_optional_text(args.schedule_expr),
+            run_at=_optional_text(args.run_at),
+            enabled=True if args.enabled is None else args.enabled,
+            model_override=args.model_override,
+            thinking_override=args.thinking_override,
+            target_idea_id=_optional_text(args.target_idea_id),
+            guidance=_optional_text(args.guidance),
+            rationale=_optional_text(args.rationale),
+        )
+        payload = serialize_cycle(cycle)
+    publish_cycle_change(action="create", **_event_from_payload(payload))
+    return {"created": payload}
+
+
+async def _action_update(ctx: ManageCycleContext) -> dict[str, Any]:
+    args = ctx.args
+    async with UnitOfWork() as uow:
+        cycle = await _load_cycle(uow.session, ctx.actor, args.id)
+        update_target_idea_id = _optional_text(args.target_idea_id)
+        if args.target_idea_id is not None:
+            await _validate_target_idea(uow.session, update_target_idea_id, ctx.actor)
+        await async_update_cycle(
+            uow.session,
+            cycle,
+            actor=ctx.actor,
+            name=_optional_text(args.name),
+            prompt=_optional_text(args.prompt),
+            timezone_name=_optional_text(args.timezone),
+            schedule_expr=_optional_text(args.schedule_expr),
+            run_at=_optional_text(args.run_at),
+            enabled=args.enabled,
+            model_override=args.model_override,
+            model_override_provided=args.model_override is not None,
+            thinking_override=_optional_text(args.thinking_override),
+            thinking_override_provided=args.thinking_override is not None,
+            target_idea_id=update_target_idea_id,
+            target_idea_provided=args.target_idea_id is not None,
+            guidance=_optional_text(args.guidance),
+            rationale=_optional_text(args.rationale),
+        )
+        payload = serialize_cycle(cycle)
+    publish_cycle_change(action="update", **_event_from_payload(payload))
+    return {"updated": payload}
+
+
+async def _action_delete(ctx: ManageCycleContext) -> dict[str, Any]:
+    async with UnitOfWork() as uow:
+        cycle = await _load_cycle(uow.session, ctx.actor, ctx.args.id)
+        await async_delete_cycle(uow.session, cycle)
+        event = cycle_change_event(cycle)
+    publish_cycle_change(action="delete", **event)
+    return {"deleted": {"id": ctx.args.id}}
+
+
+async def _action_run(ctx: ManageCycleContext) -> dict[str, Any]:
+    async with UnitOfWork() as uow:
+        cycle = await _load_cycle(uow.session, ctx.actor, ctx.args.id)
+        event = cycle_change_event(cycle)
+    payload = await async_run_cycle_now(ctx.args.id)
+    publish_cycle_change(action="run", **event)
+    return {"run": payload}
+
+
+async def _action_add_guidance(ctx: ManageCycleContext) -> dict[str, Any]:
+    async with UnitOfWork() as uow:
+        cycle = await _load_cycle(uow.session, ctx.actor, ctx.args.id)
+        row = await async_add_guidance_to_cycle(
+            uow.session,
+            cycle,
+            actor=ctx.actor,
+            guidance=_optional_text(ctx.args.guidance),
+            rationale=_optional_text(ctx.args.rationale),
+        )
+        return {"guidance": serialize_cycle_guidance(row)}
+
+
+async def _action_add_output_target(ctx: ManageCycleContext) -> dict[str, Any]:
+    async with UnitOfWork() as uow:
+        cycle = await _load_cycle(uow.session, ctx.actor, ctx.args.id)
+        row = await async_add_output_target_to_cycle(
+            uow.session,
+            cycle,
+            actor=ctx.actor,
+            target_type=_optional_text(ctx.args.output_target_type),
+            target_id=_optional_text(ctx.args.output_target_id),
+            label=_optional_text(ctx.args.output_target_label),
+            config=ctx.args.output_target_config,
+            rationale=_optional_text(ctx.args.rationale),
+        )
+        return {"output_target": serialize_cycle_output_target(row)}
+
+
+async def _action_remove_output_target(ctx: ManageCycleContext) -> dict[str, Any]:
+    output_target_id = _int_required(ctx.args.output_target_id, "output_target_id")
+    async with UnitOfWork() as uow:
+        cycle = await _load_cycle(uow.session, ctx.actor, ctx.args.id)
+        row = await async_remove_output_target_from_cycle(
+            uow.session,
+            cycle,
+            actor=ctx.actor,
+            target_id=output_target_id,
+            rationale=_optional_text(ctx.args.rationale),
+        )
+        if row is None:
+            raise ValueError(f"Cycle output target {output_target_id} not found")
+        return {"output_target": serialize_cycle_output_target(row)}
+
+
+CYCLE_ACTIONS: dict[str, CycleAction] = {
+    "list": _action_list,
+    "create": _action_create,
+    "update": _action_update,
+    "delete": _action_delete,
+    "run": _action_run,
+    "add_guidance": _action_add_guidance,
+    "add_output_target": _action_add_output_target,
+    "remove_output_target": _action_remove_output_target,
+}
+
+
+def _tool_actor(user_id: str) -> CycleMutationActor:
+    org_id = getattr(_agent_context, "org_id", None)
+    run_id = getattr(_agent_context, "run_id", None)
+    if run_id is None:
+        run = getattr(_agent_context, "run", None)
+        run_id = getattr(run, "run_id", None)
+    return CycleMutationActor(
+        user_id=user_id,
+        org_id=str(org_id) if org_id else None,
+        principal_type="agent",
+        source_id=str(run_id or user_id),
+    )
+
+
+def _cycle_scope(actor: CycleMutationActor) -> list[Any]:
+    scope = [Cycle.deleted_at.is_(None)]
+    if actor.org_id:
+        org_user_ids = select(User.id).where(User.org_id == actor.org_id)
+        scope.append(
+            or_(
+                Cycle.org_id == actor.org_id,
+                and_(Cycle.org_id.is_(None), Cycle.user_id.in_(org_user_ids)),
             )
-            mode = validate_execution_mode(execution_mode)
-            thinking = validate_thinking_override(thinking_override)
-            normalized_name = validate_nonempty_trimmed(create_name, "name")
-            normalized_prompt = validate_nonempty_trimmed(create_prompt, "prompt")
-            create_guidance = _optional_text(guidance)
-            create_rationale = _optional_text(rationale)
-            async with UnitOfWork() as uow:
-                if target_idea_id:
-                    stmt = select(Idea.id).where(*_idea_scope(target_idea_id))
-                    result = await uow.session.execute(stmt)
-                    if not result.first():
-                        return json.dumps({"error": "target_idea_id must belong to the current workspace"})
-                cycle = Cycle(
-                    user_id=user_id,
-                    org_id=org_id,
-                    creator_type="agent",
-                    creator_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                    maintainer_type="agent",
-                    maintainer_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                    name=normalized_name,
-                    prompt=normalized_prompt,
-                    schedule_expr=expr,
-                    timezone=tz_name,
-                    enabled=True if enabled is None else enabled,
-                    model_override=(model_override or "").strip() or None,
-                    thinking_override=thinking,
-                    execution_mode=REUSABLE_THREAD_EXECUTION_MODE,
-                    target_idea_id=target_idea_id,
-                    reopen_archived=cycle_defaults(
-                        execution_mode=mode,
-                        reopen_archived=reopen_archived,
-                    ),
-                    next_run_at=compute_next_run_at(expr, tz_name),
-                )
-                uow.session.add(cycle)
-                await uow.session.flush()
-                revision = await async_record_cycle_revision(
-                    uow.session,
-                    cycle,
-                    source_type="agent",
-                    source_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                    rationale=create_rationale or "Cycle created by manage_cycle.",
-                )
-                await async_add_cycle_output_target(
-                    uow.session,
-                    cycle,
-                    target_type="cycle_ledger",
-                    target_id=str(cycle.id),
-                    label="Cycle ledger",
-                    source_type="agent",
-                    source_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                    rationale="Every Cycle persists durable memory in its ledger.",
-                    revision_id=revision.id,
-                )
-                if target_idea_id:
-                    await async_add_cycle_output_target(
-                        uow.session,
-                        cycle,
-                        target_type="thread",
-                        target_id=target_idea_id,
-                        label="Cycle thread",
-                        source_type="agent",
-                        source_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                        rationale="Initial display thread for Cycle output.",
-                        revision_id=revision.id,
-                    )
-                if create_guidance:
-                    await async_add_cycle_guidance(
-                        uow.session,
-                        cycle,
-                        guidance=create_guidance,
-                        source_type="agent",
-                        source_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                        rationale=create_rationale,
-                        revision_id=revision.id,
-                    )
-                payload = serialize_cycle(cycle)
-            publish_cycle_change(
-                action="create",
-                org_id=payload.get("org_id"),
-                user_id=payload.get("user_id"),
-                cycle_id=payload.get("id"),
-                target_idea_id=payload.get("target_idea_id"),
+        )
+    else:
+        scope.append(Cycle.user_id == actor.user_id)
+    return scope
+
+
+def _idea_scope(idea_id: str, actor: CycleMutationActor) -> list[Any]:
+    scope = [Idea.id == idea_id]
+    if actor.org_id:
+        org_user_ids = select(User.id).where(User.org_id == actor.org_id)
+        scope.append(
+            or_(
+                Idea.org_id == actor.org_id,
+                and_(Idea.org_id.is_(None), Idea.user_id.in_(org_user_ids)),
             )
-            return json.dumps({"created": payload}, default=str)
-        elif action == "update":
-            if not id:
-                return json.dumps({"error": "update requires: id"})
-            update_name = _optional_update_text(name)
-            update_prompt = _optional_update_text(prompt)
-            update_timezone = _optional_update_text(timezone)
-            update_run_at = _optional_update_text(run_at)
-            update_schedule_expr = _optional_update_text(schedule_expr)
-            update_model_override = model_override if model_override is not None else None
-            update_thinking_override = _optional_update_text(thinking_override)
-            update_execution_mode = _optional_update_text(execution_mode)
-            update_target_idea_id = _optional_update_text(target_idea_id)
-            update_guidance = _optional_update_text(guidance)
-            update_rationale = _optional_update_text(rationale)
-            async with UnitOfWork() as uow:
-                stmt = select(Cycle).where(
-                    Cycle.id == id,
-                    *_cycle_scope(),
-                )
-                result = await uow.session.scalars(stmt)
-                cycle = result.first()
-                if not cycle:
-                    return json.dumps({"error": f"Cycle {id} not found"})
-                if update_target_idea_id:
-                    idea_stmt = select(Idea.id).where(*_idea_scope(update_target_idea_id))
-                    result = await uow.session.execute(idea_stmt)
-                    if not result.first():
-                        return json.dumps({"error": "target_idea_id must belong to the current workspace"})
-                if update_name is not None:
-                    cycle.name = validate_nonempty_trimmed(update_name, "name")
-                if update_prompt is not None:
-                    cycle.prompt = validate_nonempty_trimmed(update_prompt, "prompt")
-                if update_timezone is not None:
-                    cycle.timezone = validate_timezone_name(update_timezone)
-                if update_run_at is not None:
-                    cycle.schedule_expr = build_one_time_schedule_expr(update_run_at, cycle.timezone)
-                if update_schedule_expr is not None:
-                    cycle.schedule_expr = validate_schedule_expr(update_schedule_expr, cycle.timezone)
-                if enabled is not None:
-                    cycle.enabled = enabled
-                if update_model_override is not None:
-                    cycle.model_override = (update_model_override or "").strip() or None
-                if update_thinking_override is not None:
-                    cycle.thinking_override = validate_thinking_override(update_thinking_override)
-                if update_execution_mode is not None:
-                    cycle.execution_mode = validate_execution_mode(update_execution_mode)
-                if target_idea_id is not None:
-                    cycle.target_idea_id = update_target_idea_id
-                if reopen_archived is not None:
-                    cycle.reopen_archived = reopen_archived
-                elif update_execution_mode is not None:
-                    cycle.reopen_archived = cycle_defaults(
-                        execution_mode=cycle.execution_mode,
-                        reopen_archived=None,
-                    )
-                cycle.execution_mode = REUSABLE_THREAD_EXECUTION_MODE
-                cycle.reopen_archived = True
-                cycle.updated_at = datetime.now(dt_timezone.utc)
-                cycle.next_run_at = compute_next_run_at(cycle.schedule_expr, cycle.timezone)
-                revision = await async_record_cycle_revision(
-                    uow.session,
-                    cycle,
-                    source_type="agent",
-                    source_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                    rationale=update_rationale or "Cycle updated by manage_cycle.",
-                )
-                if update_guidance:
-                    await async_add_cycle_guidance(
-                        uow.session,
-                        cycle,
-                        guidance=update_guidance,
-                        source_type="agent",
-                        source_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                        rationale=update_rationale,
-                        revision_id=revision.id,
-                    )
-                payload = serialize_cycle(cycle)
-            publish_cycle_change(
-                action="update",
-                org_id=payload.get("org_id"),
-                user_id=payload.get("user_id"),
-                cycle_id=payload.get("id"),
-                target_idea_id=payload.get("target_idea_id"),
-            )
-            return json.dumps({"updated": payload}, default=str)
-        elif action == "delete":
-            if not id:
-                return json.dumps({"error": "delete requires: id"})
-            async with UnitOfWork() as uow:
-                stmt = select(Cycle).where(
-                    Cycle.id == id,
-                    *_cycle_scope(),
-                )
-                result = await uow.session.scalars(stmt)
-                cycle = result.first()
-                if not cycle:
-                    return json.dumps({"error": f"Cycle {id} not found"})
-                cycle_org_id = cycle.org_id
-                cycle_user_id = cycle.user_id
-                cycle_target_id = cycle.target_idea_id
-                cycle.enabled = False
-                cycle.deleted_at = datetime.now(dt_timezone.utc)
-            publish_cycle_change(
-                action="delete",
-                org_id=cycle_org_id,
-                user_id=cycle_user_id,
-                cycle_id=id,
-                target_idea_id=cycle_target_id,
-            )
-            return json.dumps({"deleted": {"id": id}})
-        elif action == "run":
-            if not id:
-                return json.dumps({"error": "run requires: id"})
-            async with UnitOfWork() as uow:
-                stmt = select(Cycle).where(
-                    Cycle.id == id,
-                    *_cycle_scope(),
-                )
-                result = await uow.session.scalars(stmt)
-                cycle = result.first()
-                if not cycle:
-                    return json.dumps({"error": f"Cycle {id} not found"})
-                cycle_org_id = cycle.org_id
-                cycle_user_id = cycle.user_id
-                cycle_target_id = cycle.target_idea_id
-            payload = await async_run_cycle_now(id)
-            publish_cycle_change(
-                action="run",
-                org_id=cycle_org_id,
-                user_id=cycle_user_id,
-                cycle_id=id,
-                target_idea_id=cycle_target_id,
-            )
-            return json.dumps({"run": payload}, default=str)
-        elif action == "add_guidance":
-            if not id:
-                return json.dumps({"error": "add_guidance requires: id"})
-            guidance_text = _optional_text(guidance)
-            if not guidance_text:
-                return json.dumps({"error": "add_guidance requires: guidance"})
-            async with UnitOfWork() as uow:
-                stmt = select(Cycle).where(
-                    Cycle.id == id,
-                    *_cycle_scope(),
-                )
-                result = await uow.session.scalars(stmt)
-                cycle = result.first()
-                if not cycle:
-                    return json.dumps({"error": f"Cycle {id} not found"})
-                revision = await async_record_cycle_revision(
-                    uow.session,
-                    cycle,
-                    source_type="agent",
-                    source_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                    rationale=_optional_text(rationale) or "Cycle guidance added by manage_cycle.",
-                )
-                row = await async_add_cycle_guidance(
-                    uow.session,
-                    cycle,
-                    guidance=guidance_text,
-                    source_type="agent",
-                    source_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                    rationale=_optional_text(rationale),
-                    revision_id=revision.id,
-                )
-                payload = serialize_cycle_guidance(row)
-            return json.dumps({"guidance": payload}, default=str)
-        elif action == "add_output_target":
-            if not id:
-                return json.dumps({"error": "add_output_target requires: id"})
-            target_type_text = _optional_text(output_target_type)
-            if not target_type_text:
-                return json.dumps({"error": "add_output_target requires: output_target_type"})
-            async with UnitOfWork() as uow:
-                stmt = select(Cycle).where(
-                    Cycle.id == id,
-                    *_cycle_scope(),
-                )
-                result = await uow.session.scalars(stmt)
-                cycle = result.first()
-                if not cycle:
-                    return json.dumps({"error": f"Cycle {id} not found"})
-                revision = await async_record_cycle_revision(
-                    uow.session,
-                    cycle,
-                    source_type="agent",
-                    source_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                    rationale=_optional_text(rationale) or "Cycle output target added by manage_cycle.",
-                )
-                row = await async_add_cycle_output_target(
-                    uow.session,
-                    cycle,
-                    target_type=target_type_text,
-                    target_id=_optional_text(output_target_id),
-                    label=_optional_text(output_target_label),
-                    config=output_target_config,
-                    source_type="agent",
-                    source_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                    rationale=_optional_text(rationale),
-                    revision_id=revision.id,
-                )
-                payload = serialize_cycle_output_target(row)
-            return json.dumps({"output_target": payload}, default=str)
-        elif action == "remove_output_target":
-            if not id:
-                return json.dumps({"error": "remove_output_target requires: id"})
-            target_id_text = _optional_text(output_target_id)
-            if not target_id_text:
-                return json.dumps({"error": "remove_output_target requires: output_target_id"})
-            try:
-                output_target_pk = int(target_id_text)
-            except ValueError:
-                return json.dumps({"error": "remove_output_target output_target_id must be the numeric target row id"})
-            async with UnitOfWork() as uow:
-                stmt = select(Cycle).where(
-                    Cycle.id == id,
-                    *_cycle_scope(),
-                )
-                result = await uow.session.scalars(stmt)
-                cycle = result.first()
-                if not cycle:
-                    return json.dumps({"error": f"Cycle {id} not found"})
-                revision = await async_record_cycle_revision(
-                    uow.session,
-                    cycle,
-                    source_type="agent",
-                    source_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                    rationale=_optional_text(rationale) or "Cycle output target removed by manage_cycle.",
-                )
-                row = await async_remove_cycle_output_target(
-                    uow.session,
-                    cycle,
-                    target_id=output_target_pk,
-                    source_type="agent",
-                    source_id=str(getattr(_agent_context, "run_id", None) or user_id),
-                    rationale=_optional_text(rationale),
-                    revision_id=revision.id,
-                )
-                if row is None:
-                    return json.dumps({"error": f"Cycle output target {output_target_pk} not found"})
-                payload = serialize_cycle_output_target(row)
-            return json.dumps({"output_target": payload}, default=str)
-        else:
-            return json.dumps({"error": f"Unknown action: {action}"})
-    except ValueError as e:
-        return json.dumps({"error": str(e)})
-    except Exception as e:
-        logger.exception(f"manage_cycle failed: {e}")
-        if _is_missing_cycle_schema_error(e):
-            return json.dumps(_cycle_schema_missing_payload(e))
-        return json.dumps({"error": str(e)})
+        )
+    else:
+        scope.append(Idea.user_id == actor.user_id)
+    return scope
+
+
+async def _load_cycle(session, actor: CycleMutationActor, cycle_id: int | None) -> Cycle:
+    if not cycle_id:
+        raise ValueError("id is required")
+    result = await session.scalars(select(Cycle).where(Cycle.id == cycle_id, *_cycle_scope(actor)))
+    cycle = result.first()
+    if not cycle:
+        raise ValueError(f"Cycle {cycle_id} not found")
+    return cycle
+
+
+async def _validate_target_idea(session, idea_id: str | None, actor: CycleMutationActor) -> None:
+    if not idea_id:
+        return
+    result = await session.execute(select(Idea.id).where(*_idea_scope(idea_id, actor)))
+    if not result.first():
+        raise ValueError("target_idea_id must belong to the current workspace")
+
+
+def _optional_text(value: str | None) -> str | None:
+    if value is None:
+        return None
+    text = str(value).strip()
+    return text or None
+
+
+def _int_required(value: str | None, field_name: str) -> int:
+    text = _optional_text(value)
+    if text is None:
+        raise ValueError(f"{field_name} is required")
+    try:
+        return int(text)
+    except ValueError as exc:
+        raise ValueError(f"{field_name} must be numeric") from exc
+
+
+def _event_from_payload(payload: dict[str, Any]) -> dict[str, Any]:
+    return {
+        "org_id": payload.get("org_id"),
+        "user_id": payload.get("user_id"),
+        "cycle_id": payload.get("id"),
+        "target_idea_id": payload.get("target_idea_id"),
+    }
+
 
 __all__ = [name for name in globals() if not name.startswith("__")]
